refactor(settings): split monolith into 8-group left-nav layout (#57)
The 1573-line /settings/page.tsx with 13 stacked cards had grown past
the size where users could find anything. Split into 8 grouped sub-pages
(General, Account & Security, Data, Categorization, Investments,
Display & Ordering, Integrations, Developer) under a shared left-nav
shell. The bare /settings URL now server-side redirects to
/settings/general so every existing link keeps working.

Pure UX rework — every card calls the same APIs with the same payloads,
same useState shape, same handlers, same fetch payloads. Categorization
ended up at 585 lines (vs spec target of ~400) because Category
Management + Transaction Rules form one coherent UX group; rest fit
under the target. 1573 -> largest sub-page 585 (63% reduction).

Closes #57
